Feature/488 production seeders

+++ src/seeders/20200503190707-development-add-users.ts
@@ -2,9 +2,11 @@ import { QueryInterface } from "sequelize";
 import { dylan, manuel, sebastian } from "./constants/applicants";
 import { claudio, marcos } from "./constants/companyUsers";
 import { federico, aldana } from "./constants/admins";
+import { Environment } from "../config/Environment";
 
 export = {
-  up: (queryInterface: QueryInterface) => {
+  up: async (queryInterface: QueryInterface) => {
+    if (Environment.NODE_ENV() === Environment.PRODUCTION) return;
     return queryInterface.bulkInsert("Users", [
       dylan.user,
       manuel.user,
@@ -15,7 +17,8 @@ export = {
       aldana.user
     ]);
   },
-  down: (queryInterface: QueryInterface) => {
+  down: async (queryInterface: QueryInterface) => {
+    if (Environment.NODE_ENV() === Environment.PRODUCTION) return;
     return queryInterface.bulkDelete("Users", {});
   }
 };
