+++ src/seeders/20201226014356-development-add-shared-settings.ts
@@ -1,7 +1,9 @@
 import { QueryInterface } from "sequelize";
+import { Environment } from "../config/Environment";
 
 export = {
-  up: (queryInterface: QueryInterface) => {
+  up: async (queryInterface: QueryInterface) => {
+    if (Environment.NODE_ENV() === Environment.PRODUCTION) return;
     return queryInterface.bulkInsert("SharedSettings", [
       {
         uuid: "6b228e77-9e8e-4438-872e-f3714a5846dd",
@@ -14,7 +16,8 @@ export = {
       }
     ]);
   },
-  down: (queryInterface: QueryInterface) => {
+  down: async (queryInterface: QueryInterface) => {
+    if (Environment.NODE_ENV() === Environment.PRODUCTION) return;
     return queryInterface.bulkDelete("SharedSettings", {});
   }
 };
